Rewritten ScreenCapture to directly save as UINT16. Renamed both Depth and BGR output files (.depth16, .bgr8)

+++ Source/RTW_Simulation/CustomScreenCapture.h
@@ -63,7 +63,7 @@ protected:
 	class USceneCaptureComponent2D* sceneCaptureColor;
 	class UCameraComponent* OurCameraColor;
 
-	std::vector<float> depthVector;
+	std::vector<uint16_t> depthVector;
 
 	// Called when the game starts or when spawned
 	virtual void BeginPlay() override;
